[CVE] Only consider bounded matchers, fix tor issues
Sometimes the NVD contains unbounded matchers that match everything
for example https://nvd.nist.gov/vuln/detail/CVE-2009-0414 has a
matcher of

 cpe:2.3:a:tor:tor:*:*:*:*:*:*:*:*

without any bounds. Lars and I decided to ignore these CPE matches
because it seems nonsensical or at least not useful for there to be a
CVE that cannot be fixed.