Refactor the update function to make it easier to add rewriters; add rewriter for meta.homepage
Per discussion, we are interested in evolving `nixpkgs-update` to be a bot that
slowly cleans up arbitrary bits of "legacy" nix code as part of the standard
package version update process.

This does some rewriting of the main Update function to factor out the existing
version-bumper as a rewrite function, and add a second one for updating
meta.homepage to be quoted (if not done already).