Minor bugfix for quoted urls when homepage is not in the file
The rewriter would do the right thing, but the comment would still show up in
the PR description. This adds a check to the quotedUrls rewriter to handle the
case where the homepage DOES exist and is NOT quoted, but also would not be it
by the text replace, because it inherits it from another file.