Just print the diff once at the end of rewrites
Particularly as we add more specific rewrite functions, the norm is that most of
them will not generate a diff on a particular update, so the in-between diff
should be mostly the same. This keeps the log clean by just printing the final
diff; since we have msgs for rewrites fired we can still (somewhat) investigate
what did what when things go wrong.