Refactor rewriters to reduce duplication
This commit moves the meat of the standard `src` and `version` update into a
helper function that can be re-used across the regular version updater and the
Rust package updater.

Since I am soon to add it to a third place for updating Golang packages, which
structurally look a lot like Rust updates, I think it makes sense to move this
to a helper function rather than copy it in a third place.

Also adds a newline in front of the section on `Updates Performed` to make the
PR comment prettier.

All in all, no functional changes intended, just cleanups/prep for more features.