Automated updates for buildGoModule packages
Having merged an update to make go packages passthru the `modSha256` [1], we can
now have `nixpkgs-update` inspect it and make updates, just as it now does for
Rust packages as of [2].

As a slight detail, because the attribute is in `passthru` rather than actually
being available in the derivation, we just do the equivalent of
`nix eval -f . <attrPath>.<attr>` rather than going through the `drvAttrs` as before.

This version has been tested successfully with at least one go update [3], but
not thoroughly vetted. That said, it's essentially identical to the Rust
implementation that has been working well for some time now.

[1] https://github.com/NixOS/nixpkgs/pull/82027
[2] https://github.com/ryantm/nixpkgs-update/pull/156
[3] https://github.com/NixOS/nixpkgs/pull/82465