Cleanup whitespace in PR comment function
Since we have an interpolated string, let's try to keep the newlines inside the
interpolation region, rather than attached to the variables -- particularly when
it's multiple newlines. This makes it easier to read/visualize what the PR
description will look like.

Updated and trimmed some excessive whitespace. This doesn't cause a rendering
diff on the UI, but makes the raw output cleaner.