Add workarounds for Firefox binary packages
The binary Firefox packages don't get cached in Hydra and don't have
their version numbers appear as strings in the file that `nix edit`
identifies. To solve these problems, we add two escape hatches: a list
of packages for which outpath diffing is skipped, and a list of
overrides to use instead of `nix edit` for identifying the correct Nix
file to scan for version numbers.