support .originalSrc attr as a fallback
If a builder creates a sequence of derivations on the way to
building a package, nixpkgs-update won't be able to find the
correct src attr to query. We can technically use passthru
to lie about the src, but src is load-bearing and this may
mislead or confuse humans (or other tools?) later. This also
checks an originalSrc fallback attr that isn't load-bearing.