Gpg: init
This patch creates a new Gpg module which will be used to verify
signatures of signed releases. This will allow nixpkgs-update to warn
whenever a release is detected to be made by an unknown party - in which
case the package maintainers should verify whether the key used to sign
the release is made by upstream maintainers.